Fold new util file back into general file
