+++ include/battle_controllers.h
@@ -340,4 +340,13 @@ void SetControllerToLinkOpponent(u32 battler);
 // link partner
 void SetControllerToLinkPartner(u32 battler);
 
+void TrySetBattlerShadowSpriteCallback(u32 battler);
+
+bool32 TryShinyAnimAfterMonAnimUtil(u32 battler);
+bool32 SwitchIn_ShowSubstituteUtil(u32 battler, bool32 isPlayerSide);
+bool32 SwitchIn_WaitAndEndUtil(u32 battler);
+bool32 SwitchIn_HandleSoundAndEndUtil(u32 battler, bool32 isPlayerSide);
+bool32 SwitchIn_ShowHealthboxUtil(u32 battler, bool32 isPlayerSide);
+bool32 SwitchIn_TryShinyAnimUtil(u32 battler, bool32 isPlayerSide);
+
 #endif // GUARD_BATTLE_CONTROLLERS_H
