+++ src/battle_controller_link_partner.c
@@ -3,7 +3,6 @@
 #include "battle_ai_main.h"
 #include "battle_anim.h"
 #include "battle_controllers.h"
-#include "battle_controller_util.h"
 #include "battle_interface.h"
 #include "battle_message.h"
 #include "battle_setup.h"
