+++ src/battle_controller_opponent.c
@@ -6,7 +6,6 @@
 #include "battle_anim.h"
 #include "battle_arena.h"
 #include "battle_controllers.h"
-#include "battle_controller_util.h"
 #include "battle_message.h"
 #include "battle_interface.h"
 #include "battle_setup.h"
