+++ src/battle_controller_player_partner.c
@@ -4,7 +4,6 @@
 #include "battle_ai_util.h"
 #include "battle_anim.h"
 #include "battle_controllers.h"
-#include "battle_controller_util.h"
 #include "battle_message.h"
 #include "battle_interface.h"
 #include "battle_setup.h"
